Restructuring lib
