+++ src/lib/FSM/Event/Event.cpp
@@ -1,0 +1,21 @@
+//
+// Created by Michael Fields on 2/21/21.
+//
+
+#include "Event.h"
+
+using namespace std;
+
+Event::Event(string name, string type ) {
+    this->Type = type;
+    this->Name = name;
+    this->Timing = 0;
+    this->executed = false;
+}
+
+Event::Event(string name, string type, long timing)  {
+    this->Type = type;
+    this->Name = name;
+    this->Timing = timing;
+    this->executed = false;
+}
